Moving apollo client init to seperate file.
